requirements.txt
absl-py==2.4.0
altair==5.5.0
asttokens==3.0.1
astunparse==1.6.3
attrs==26.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.6
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
flatbuffers==25.12.19
fonttools==4.62.1
gast==0.4.0
gitdb==4.0.12
gitpython==3.1.46
google-auth==2.49.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.80.0
h5py==3.16.0
idna==3.11
ipykernel==7.2.0
ipython==8.39.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
keras==2.10.0
keras-preprocessing==1.1.2
keras-tuner==1.4.8
kiwisolver==1.5.0
kt-legacy==1.0.5
libclang==18.1.1
markdown==3.10.2
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
narwhals==2.18.1
nest-asyncio==1.6.0
numpy==1.23.5
oauthlib==3.3.1
opt-einsum==3.4.0
packaging==23.2
pandas==2.3.3
parso==0.8.6
pillow==10.4.0
platformdirs==4.9.4
prompt-toolkit==3.0.52
protobuf==3.20.3
psutil==7.2.2
pure-eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydeck==0.9.1
pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
pyzmq==27.1.0
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.15.3
setuptools==82.0.1
six==1.17.0
smmap==5.0.3
stack-data==0.6.3
streamlit==1.32.0
tenacity==8.5.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
traitlets==5.14.3
typing-extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.6.0
werkzeug==3.1.8
wheel==0.46.3
wrapt==2.1.2
